si no jala bye 5

+++ src/main/java/com/upiiz/platform_api/repositories/AdminTopicInterestProjection.java
@@ -9,7 +9,7 @@ public interface AdminTopicInterestProjection {
     Long getSubareaId();
     String getSubareaName();
     Long getTotalEvents();
-    Integer getWeightedScore();
+    Long getWeightedScore();
     Long getUniqueUsers();
     LocalDateTime getLastEventAt();
 }
